backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.18
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
python-dotenv==1.1.0
starlette==0.41.3
setuptools==80.9.0